refactor: moved command_builder
